Migrate the Enzyme test __tests__ root directory and remove __tests__ directory
